+++ client/v2/autocli/flag/builder.go
@@ -27,6 +27,7 @@ const (
 	AddressStringScalarType          = "cosmos.AddressString"
 	ValidatorAddressStringScalarType = "cosmos.ValidatorAddressString"
 	ConsensusAddressStringScalarType = "cosmos.ConsensusAddressString"
+	PubkeyScalarType                 = "cosmos.Pubkey"
 )
 
 // Builder manages options for building pflag flags for protobuf messages.
@@ -67,6 +68,7 @@ func (b *Builder) init() {
 		b.scalarFlagTypes[AddressStringScalarType] = addressStringType{}
 		b.scalarFlagTypes[ValidatorAddressStringScalarType] = validatorAddressStringType{}
 		b.scalarFlagTypes[ConsensusAddressStringScalarType] = consensusAddressStringType{}
+		b.scalarFlagTypes[PubkeyScalarType] = pubkeyType{}
 	}
 }